Fix console warning
